actualizacion datos

+++ Prog. Orientada a Objetos/02-Gestion 1º DAW/Gestion/Models/DatosAcademicos.cs
@@ -5,5 +5,5 @@
 public struct DatosAcademicos {
     public int Faltas { get; set; }
     public int Retrasos { get; set; }
-    public int NotaProgramacion { get; set; }
+    public double NotaProgramacion { get; set; }
 }
